requirements.txt
anyio==4.9.0
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.1.0
python-telegram-bot==22.0
pytubefix==8.12.3
sniffio==1.3.1
